+++ test_alconna/entry_test.py
@@ -8,6 +8,6 @@ from test_alconna.components_test import *
 from test_alconna.config_test import *
 from test_alconna.analyser_test import *
 
-
 if __name__ == '__main__':
+    import pytest
     pytest.main([__file__, "-vs"])